Implemented DE-695

+++ tutorial/src/main/java/com/arangodb/spring/demo/runner/CrudRunner.java
@@ -20,16 +20,23 @@
 
 package com.arangodb.spring.demo.runner;
 
+import com.arangodb.model.TtlIndexOptions;
 import com.arangodb.spring.demo.entity.Character;
 import com.arangodb.spring.demo.repository.CharacterRepository;
 import com.arangodb.springframework.core.ArangoOperations;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.stereotype.Component;
+
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.Collections;
+import java.util.Date;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ComponentScan("com.arangodb.spring.demo")
+@Component
 public class CrudRunner implements CommandLineRunner {
 
     @Autowired
@@ -38,21 +45,41 @@ public class CrudRunner implements CommandLineRunner {
     @Autowired
     private CharacterRepository repository;
 
+    @Value("${arangodb.ddl.enabled:false}")
+    private boolean isDdlEnabled;
+
     @Override
     public void run(String... args) {
-        // first drop the database so that we can run this multiple times with the same dataset
-        operations.dropDatabase();
+        if (isDdlEnabled) {
+            operations.dropDatabase();
+            System.out.println("Database dropped for fresh DDL operations.");
+            createTTLIndex();
 
-        System.out.println("# CRUD operations");
+            System.out.println("Schema validation passed.");
 
-        // save a single entity in the database
-        // there is no need of creating the collection first. This happen automatically
-        Character nedStark = new Character(null, "Ned", "Stark");
-        Character saved = repository.save(nedStark);
-        System.out.println("Ned Stark saved in the database: " + saved);
+            // Set document expiration for 2 minutes from now
+            Long expiresAt = Instant.now().plus(2, ChronoUnit.MINUTES).getEpochSecond();
+            Character nedStark = new Character(null, "John", "Williams", expiresAt);
+            Character saved = repository.save(nedStark);
 
-        assertThat(saved.id()).isNotNull();
-        assertThat(saved.name()).isEqualTo(nedStark.name());
-        assertThat(saved.surname()).isEqualTo(nedStark.surname());
+            System.out.println("Saved character: " + saved);
+            System.out.println("Document expires at: " + new Date(saved.expiresAt() * 1000));
+            System.out.println("Current time: " + new Date());
+
+            assertThat(saved.id()).isNotNull();
+            assertThat(saved.name()).isEqualTo(nedStark.name());
+            assertThat(saved.surname()).isEqualTo(nedStark.surname());
+        } else {
+            System.out.println("DDL operations are disabled. Skipping schema creation.");
+            System.out.println(
+                    "You can enable DDL operations by setting 'arangodb.ddl.enabled=true' in application.properties.");
+        }
+    }
+    private void createTTLIndex() {
+        System.out.println("Creating TTL index...");
+        operations.collection(Character.class).ensureTtlIndex(
+                Collections.singletonList("expiresAt"),
+                new TtlIndexOptions().expireAfter(0));
+        System.out.println(" TTL index created on 'expiresAt'.");
     }
 }
